requirements.txt
requests
pycryptodome
# 北理贝塔驿站浏览 skill 依赖
